requirements.txt
garmy
python-dotenv
requests